requirements.txt
aws-cdk-lib==2.202.0
constructs>=10.0.0,<11.0.0
pytest==6.2.5
bayoo-docx>=0.2.1
redlines>=0.4.2
boto3==1.34.0
